Move ground station to its own module for maintainability
